requirement.txt
sentence-transformers
flask
numpy
python-dotenv